Add additional properties
