+++ src/types.ts
@@ -1,12 +1,14 @@
 interface HappyHour {
-  day: string;
-  startTime: string;
-  endTime: string;
+  day: number;
+  startTime: number;
+  endTime: number;
 }
 
 export interface Pub {
   id: number;
   name: string;
+  website: string;
+  logo: string;
   coordinate: {
     latitude: number;
     longitude: number;
